LPD-39873 Manual Processing email template trigger
